Enable PostgreSQL and run `pgagroal-cli status` in CI
This commit modifies the `build-linux` and `build-macos` workflows
by incorporating PostgreSQL and by testing pgagroal for both
GCC and CLANG builds. The changes ensure that the workflow:

(a) Installs and runs the latest version of PostgreSQL;
(b) Runs pgagroal; and
(c) Runs `pgagroal-cli status` command and to confirm that
pgagroal is operational.

For the macOS build, the `brew install postgresql@NN` command is run
with --overwrite flag in order to force Homebrew to overwrite
conflicting files during the installation.