Enable PostgreSQL and run `pgagroal-cli status` in CI
This commit modifies the `build-linux` and `build-macos` workflows
by incorporating PostgreSQL and by testing pgagroal for both
GCC and CLANG builds. The changes ensure that the workflow:

(a) Installs and runs the latest version of PostgreSQL;
(b) Runs pgagroal; and
(c) Runs `pgagroal-cli status` command and to confirm that
pgagroal is operational.

For the macOS build, the `brew install postgresql@NN` command fails
due to conflicts during the creation of links for dependencies,
although it successfully installs postgresql.

A more sophisticated script could solve this issue, but ignoring the
errors with `|| true` serves as a practical workaround.