Enable PostgreSQL and run pgagroal-cli status in CI
This commit modifies the build-linux workflow by incorporating
PostgreSQL and by testing pgagroal for both GCC and CLANG build.
The changes ensure that the workflow:

(a) Installs and runs the latest version of PostgreSQL;
(b) Runs pgagroal; and
(c) Runs pgagroal-cli status command and to confirm that
pgagroal's status is "Running".